set block user report delete and zoom

+++ app/src/main/java/com/stormstreet/myapplication/BottomSheets/BottomSheetOptions.kt
@@ -11,7 +11,7 @@ import com.stormstreet.myapplication.customclickListner.ClickListnerDelete
 import com.stormstreet.myapplication.entity.Service_Base.ServiceManager
 import com.google.android.material.bottomsheet.BottomSheetDialogFragment
 
-class BottomSheetOptions(var click : ClickListnerDelete) : BottomSheetDialogFragment(){
+class BottomSheetOptions(var click: ClickListnerDelete,var tagdata: String) : BottomSheetDialogFragment(){
     lateinit var mContext : Context
     lateinit var serviceManager: ServiceManager
     lateinit var deletepost: TextView
@@ -25,15 +25,42 @@ class BottomSheetOptions(var click : ClickListnerDelete) : BottomSheetDialogFrag
         var v = inflater.inflate(R.layout.bottom_options, container, false)
         mContext = activity!!.applicationContext
         serviceManager = ServiceManager(mContext)
+
         deletepost = v.findViewById(R.id.deletepost)
         cancel = v.findViewById(R.id.cancel)
-
+//=========================
+        if(tagdata.equals("delete"))
+        {
+            deletepost.setText("Delete Post")
+        }
+        else
+        if(tagdata.equals("block"))
+        {
+            deletepost.setText("Block User")
+        }
+        else
+        {
+            deletepost.setText("Report")
+        }
 
         cancel.setOnClickListener {
             dismiss()
         }
         deletepost.setOnClickListener {
-          click.deletePost()
+            if(tagdata!!.equals("delete"))
+            {
+                click.deletePost()
+            }
+            else
+                if(tagdata.equals("block"))
+                {
+                    click.deletePost()
+                }
+            else
+            {
+                click.reportpost()
+            }
+
         }
         return v
     }
